requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.1
django-allauth==0.47.0
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
Pillow==9.0.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.5
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==5.3.0
